# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ECHO Anonymous Messenger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  username: string;
  publicIdentityKey: string;
  publicSignedPrekey?: string;
  publicOneTimePrekeys?: string[];
  seedHash: string;
}

export interface LoginInput {
  username: string;
  seedHash: string;
  publicDeviceKey?: string;
  signature?: string;
}

export interface RestoreInput {
  seedHash: string;
  publicDeviceKey?: string;
  newPublicIdentityKey?: string;
}

export interface DeleteAccountInput {
  seedHash: string;
  signature?: string;
}

export interface AuthResult {
  userId: number;
  username: string;
  sessionToken: string;
  deviceId: string;
}

export interface UsernameCheck {
  username: string;
}

export interface UsernameAvailability {
  available: boolean;
}

export interface PublicUser {
  id: number;
  username: string;
  publicIdentityKey: string;
  publicSignedPrekey?: string;
  /** @nullable */
  oneTimePrekey?: string | null;
  /** @nullable */
  lastActive?: string | null;
  /** @nullable */
  avatarFileId?: string | null;
}

export interface UserSettings {
  theme?: string;
  language?: string;
  notifications?: boolean;
  showLastSeen?: boolean;
  showAvatar?: boolean;
  /** @nullable */
  autoDeleteMessages?: number | null;
}

export interface UserProfile {
  id: number;
  username: string;
  publicIdentityKey: string;
  lastActive: string;
  /** @nullable */
  avatarFileId?: string | null;
  settings?: UserSettings;
  createdAt: string;
}

export interface SettingsUpdate {
  settings?: UserSettings;
}

export interface PrekeysInput {
  publicOneTimePrekeys: string[];
}

export interface Message {
  id: number;
  senderId: number;
  senderUsername?: string;
  chatId: number;
  chatType: number;
  encryptedContent: string;
  timestamp: string;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  readAt?: string | null;
  deleteAfterRead: boolean;
  /** @nullable */
  deleteAfterSeconds?: number | null;
  isVoice: boolean;
  /** @nullable */
  mediaFileId?: string | null;
  /** @nullable */
  replyToId?: number | null;
}

export interface MessageInput {
  /** @nullable */
  receiverId?: number | null;
  /** @nullable */
  chatId?: number | null;
  chatType: number;
  encryptedContent: string;
  deleteAfterRead?: boolean;
  /** @nullable */
  deleteAfterSeconds?: number | null;
  isVoice?: boolean;
  /** @nullable */
  mediaFileId?: string | null;
  /** @nullable */
  replyToId?: number | null;
}

export interface ChatSummary {
  id: number;
  type: number;
  title: string;
  /** @nullable */
  avatarFileId?: string | null;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount: number;
  isPinned: boolean;
  memberCount: number;
  isSecret: boolean;
}

export type ChatSettings = { [key: string]: unknown };

export interface Chat {
  id: number;
  type: number;
  title: string;
  creatorId: number;
  /** @nullable */
  avatarFileId?: string | null;
  createdAt: string;
  memberCount: number;
  settings?: ChatSettings;
}

export interface ChatInput {
  type: number;
  title: string;
  memberIds?: number[];
}

export type ChatUpdateSettings = { [key: string]: unknown };

export interface ChatUpdate {
  title?: string;
  /** @nullable */
  avatarFileId?: string | null;
  settings?: ChatUpdateSettings;
}

export interface ChatMember {
  userId: number;
  username: string;
  role: string;
  joinedAt: string;
  /** @nullable */
  avatarFileId?: string | null;
}

export interface AddMemberInput {
  userId: number;
  role?: string;
}

export interface ChatStats {
  totalChats: number;
  totalMessages: number;
  totalContacts: number;
  secretChats: number;
  unreadTotal: number;
}

export interface FileUploadInput {
  /** Base64-encoded encrypted file data */
  data: string;
  mimeType?: string;
  encryptedKey?: string;
}

export interface FileUploadResult {
  fileId: string;
  url: string;
  size: number;
  mimeType?: string;
}

export interface FileInfo {
  fileId: string;
  size: number;
  mimeType?: string;
  createdAt: string;
}

export interface Session {
  deviceId: string;
  lastUsed: string;
  createdAt: string;
  userAgent?: string;
  isCurrent: boolean;
}

export interface BackupInput {
  encryptedData: string;
}

export interface BackupInfo {
  id: string;
  encryptedData?: string;
  createdAt: string;
}

export type GetMessagesParams = {
chatId: number;
before?: number;
limit?: number;
};