# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ECHO Anonymous Messenger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new anonymous user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "409":
          description: Username taken
  /check-username:
    post:
      operationId: checkUsername
      tags: [auth]
      summary: Check username availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsernameCheck"
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameAvailability"
  /login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with seed hash and device key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /restore:
    post:
      operationId: restoreAccount
      tags: [auth]
      summary: Restore account via seed hash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreInput"
      responses:
        "200":
          description: Account restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid seed
  /delete-account:
    post:
      operationId: deleteAccount
      tags: [auth]
      summary: Burn account - delete all data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountInput"
      responses:
        "200":
          description: Account burned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /chats:
    get:
      operationId: getChats
      tags: [chats]
      summary: Get all chats for current user
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSummary"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a group or channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "201":
          description: Chat created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /chats/{id}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get chat details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
    patch:
      operationId: updateChat
      tags: [chats]
      summary: Update chat settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatUpdate"
      responses:
        "200":
          description: Chat updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /chats/{id}/members:
    get:
      operationId: getChatMembers
      tags: [chats]
      summary: Get members of a group/channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMember"
    post:
      operationId: addChatMember
      tags: [chats]
      summary: Add member to chat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberInput"
      responses:
        "200":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /chats/{id}/members/{userId}:
    delete:
      operationId: removeChatMember
      tags: [chats]
      summary: Remove member from chat
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /sessions/{deviceId}:
    delete:
      operationId: terminateSession
      tags: [auth]
      summary: Terminate a session
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /backup/upload:
    post:
      operationId: uploadBackup
      tags: [auth]
      summary: Upload encrypted backup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupInput"
      responses:
        "200":
          description: Backup uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    RegisterInput:
      type: object
      properties:
        username:
          type: string
        publicIdentityKey:
          type: string
        publicSignedPrekey:
          type: string
        publicOneTimePrekeys:
          type: array
          items:
            type: string
        seedHash:
          type: string
      required: [username, publicIdentityKey, seedHash]
    LoginInput:
      type: object
      properties:
        username:
          type: string
        seedHash:
          type: string
        publicDeviceKey:
          type: string
        signature:
          type: string
      required: [username, seedHash]
    RestoreInput:
      type: object
      properties:
        seedHash:
          type: string
        publicDeviceKey:
          type: string
        newPublicIdentityKey:
          type: string
      required: [seedHash]
    DeleteAccountInput:
      type: object
      properties:
        seedHash:
          type: string
        signature:
          type: string
      required: [seedHash]
    AuthResult:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        sessionToken:
          type: string
        deviceId:
          type: string
      required: [userId, username, sessionToken, deviceId]
    UsernameCheck:
      type: object
      properties:
        username:
          type: string
      required: [username]
    UsernameAvailability:
      type: object
      properties:
        available:
          type: boolean
      required: [available]
...
    ChatSummary:
      type: object
      properties:
        id:
          type: integer
        type:
          type: integer
        title:
          type: string
        avatarFileId:
          type: ["string", "null"]
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        unreadCount:
          type: integer
        isPinned:
          type: boolean
        memberCount:
          type: integer
        isSecret:
          type: boolean
      required: [id, type, title, unreadCount, isPinned, memberCount, isSecret]
    Chat:
      type: object
      properties:
        id:
          type: integer
        type:
          type: integer
        title:
          type: string
        creatorId:
          type: integer
        avatarFileId:
          type: ["string", "null"]
        createdAt:
          type: string
        memberCount:
          type: integer
        settings:
          type: object
      required: [id, type, title, creatorId, createdAt, memberCount]
    ChatInput:
      type: object
      properties:
        type:
          type: integer
        title:
          type: string
        memberIds:
          type: array
          items:
            type: integer
      required: [type, title]
    ChatUpdate:
      type: object
      properties:
        title:
          type: string
        avatarFileId:
          type: ["string", "null"]
        settings:
          type: object
    ChatMember:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        role:
          type: string
        joinedAt:
          type: string
        avatarFileId:
          type: ["string", "null"]
      required: [userId, username, role, joinedAt]
    AddMemberInput:
      type: object
      properties:
        userId:
          type: integer
        role:
          type: string
      required: [userId]
...
    BackupInput:
      type: object
      properties:
        encryptedData:
          type: string
      required: [encryptedData]
...