# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ECHO Anonymous Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthResult {
  userId: number;
  username: string;
  sessionToken: string;
  deviceId: string;
}