# lib/api-zod/src/generated/types/chatSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ECHO Anonymous Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatSummary {
  id: number;
  type: number;
  title: string;
  /** @nullable */
  avatarFileId?: string | null;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount: number;
  isPinned: boolean;
  memberCount: number;
  isSecret: boolean;
}