# lib/api-zod/src/generated/types/restoreInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ECHO Anonymous Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface RestoreInput {
  seedHash: string;
  publicDeviceKey?: string;
  newPublicIdentityKey?: string;
}